integration tests now using only 1 container
