+++ backend/uasmartsignage/src/main/resources/.testcontainers.properties
@@ -1,0 +1,1 @@
+testcontainers.reuse.enable=true
