+++ backend/uasmartsignage/src/test/java/deti/uas/uasmartsignage/integrationTests/BaseIntegrationTest.java
@@ -1,0 +1,38 @@
+package deti.uas.uasmartsignage.integrationTests;
+
+import org.junit.After;
+import org.junit.jupiter.api.*;
+import org.junit.jupiter.api.TestInstance.Lifecycle;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.boot.test.context.SpringBootTest.WebEnvironment;
+import org.springframework.boot.test.web.server.LocalServerPort;
+import org.springframework.boot.test.web.client.TestRestTemplate;
+import org.springframework.test.context.ContextConfiguration;
+import org.springframework.test.context.DynamicPropertyRegistry;
+import org.springframework.test.context.DynamicPropertySource;
+import org.testcontainers.containers.PostgreSQLContainer;
+import org.testcontainers.junit.jupiter.Container;
+import org.testcontainers.junit.jupiter.Testcontainers;
+
+@SpringBootTest(webEnvironment = WebEnvironment.RANDOM_PORT, properties = {"spring.profiles.active=test"})
+@TestMethodOrder(MethodOrderer.OrderAnnotation.class)
+public abstract class BaseIntegrationTest {
+
+    public static PostgreSQLContainer container;
+
+    static {
+        container = (PostgreSQLContainer) new PostgreSQLContainer<>("postgres:latest").withReuse(true)
+                .withDatabaseName("uasmartsignageIT")
+                .withUsername("integrationTest")
+                .withPassword("test");
+        container.start();
+    }
+
+    @DynamicPropertySource
+    static void properties(DynamicPropertyRegistry registry) {
+        registry.add("spring.datasource.url", container::getJdbcUrl);
+        registry.add("spring.datasource.username", container::getUsername);
+        registry.add("spring.datasource.password", container::getPassword);
+    }
+}
